refactor(minecraft): extract library management into ModLibrary
